Move EventConferencing to a subclass of Event
