+++ src/main/java/com/nylas/Event.java
@@ -25,7 +25,7 @@ public class Event extends AccountOwnedModel implements JsonObject {
 	private Boolean read_only;
 	private Boolean busy;
 	private Map<String, String> metadata;
-	private EventConferencing conferencing;
+	private Conferencing conferencing;
 	
 	private Recurrence recurrence;
 	
@@ -92,7 +92,7 @@ public class Event extends AccountOwnedModel implements JsonObject {
 		return metadata;
 	}
 
-	public EventConferencing getConferencing() {
+	public Conferencing getConferencing() {
 		return conferencing;
 	}
 
@@ -149,7 +149,7 @@ public class Event extends AccountOwnedModel implements JsonObject {
 		this.metadata = metadata;
 	}
 
-	public void setConferencing(EventConferencing conferencing) {
+	public void setConferencing(Conferencing conferencing) {
 		this.conferencing = conferencing;
 	}
 
@@ -199,7 +199,89 @@ public class Event extends AccountOwnedModel implements JsonObject {
 			return "Recurrence [timezone=" + timezone + ", rrule=" + rrule + "]";
 		}
 	}
-	
+
+	public static class Conferencing {
+		private String provider;
+		private Details details;
+
+		/** For deserialiation only */ public Conferencing() {}
+
+		public String getProvider() {
+			return provider;
+		}
+
+		public void setProvider(String provider) {
+			this.provider = provider;
+		}
+
+		public Details getDetails() {
+			return details;
+		}
+
+		public void setDetails(Details details) {
+			this.details = details;
+		}
+
+		@Override
+		public String toString() {
+			return "Conferencing [provider=" + provider + ", details=" + details + "]";
+		}
+
+		public static class Details {
+			private String url;
+			private String password;
+			private String pin;
+			private String meeting_code;
+			private List<String> phone;
+
+			public String getUrl() {
+				return url;
+			}
+
+			public void setUrl(String url) {
+				this.url = url;
+			}
+
+			public String getPassword() {
+				return password;
+			}
+
+			public void setPassword(String password) {
+				this.password = password;
+			}
+
+			public String getPin() {
+				return pin;
+			}
+
+			public void setPin(String pin) {
+				this.pin = pin;
+			}
+
+			public String getMeetingCode() {
+				return meeting_code;
+			}
+
+			public void setMeetingCode(String meetingCode) {
+				this.meeting_code = meetingCode;
+			}
+
+			public List<String> getPhone() {
+				return phone;
+			}
+
+			public void setPhone(List<String> phone) {
+				this.phone = phone;
+			}
+
+			@Override
+			public String toString() {
+				return String.format("Details [url=%s, password=%s, pin=%s, meeting_code=%s, phone=%s]",
+						url, password, pin, meeting_code, phone);
+			}
+		}
+	}
+
 	public static interface When extends JsonObject {}
 	
 	/**
